implement application repository

+++ src/services/application.service.ts
@@ -4,6 +4,8 @@ import { APPLICATION_MONGO_DB_REPOSITORY, httpError } from "../utils/const";
 import { ServiceError } from "../utils/service-errors";
 import type Context from "../middlewares/context.ts";
 import type { IApplicationRepository } from "../repositories/interfaces.ts";
+import { type Environment } from "../models/environment.model.ts";
+import { PrismaClientKnownRequestError } from "@prisma/client/runtime/library";
 
 const getApplication = async ({
   context,
@@ -13,7 +15,11 @@ const getApplication = async ({
   context: Context;
   appName: string;
   clerkId: string;
-}): Promise<Application> => {
+}): Promise<
+  Omit<Application, "environments"> & {
+    environments: Pick<Environment, "_id" | "name" | "description">[];
+  }
+> => {
   const repository = context.get<IApplicationRepository>(
     APPLICATION_MONGO_DB_REPOSITORY,
   );
@@ -64,7 +70,9 @@ const createApplication = async ({
       appDescription,
     });
   } catch (e: any) {
-    if (e.code === 11000) {
+    if (e instanceof PrismaClientKnownRequestError && e.code === "P2002") {
+      throw new ServiceError(httpError.APPNAME_EXISTS);
+    } else if (e.code === 11000) {
       throw new ServiceError(httpError.APPNAME_EXISTS);
     } else {
       console.log("Error creating app", e);
@@ -80,7 +88,7 @@ const updateApplication = async (props: {
   clerkId: string;
   description?: string;
   image?: string;
-}): Promise<Application | null> => {
+}): Promise<Omit<Application, "environments"> | null> => {
   const repository = props.context.get<IApplicationRepository>(
     APPLICATION_MONGO_DB_REPOSITORY,
   );
@@ -92,14 +100,15 @@ const updateApplication = async (props: {
   }) as { name?: string; description?: string; image?: string };
 
   try {
-    const result = await repository.updateApplication({
+    return repository.updateApplication({
       oldAppName: props.oldAppName,
       clerkId: props.clerkId,
       updateProps,
     });
-
-    return result;
   } catch (e) {
+    if (e instanceof PrismaClientKnownRequestError && e.code === "P2001") {
+      return null; // app does not exist prisma error
+    }
     console.log("Error updating app", e);
     throw new ServiceError(httpError.UNKNOWN);
   }
@@ -113,10 +122,15 @@ const deleteApplication = async ({
   context: Context;
   appName: string;
   clerkId: string;
-}): Promise<Application | null> => {
+}): Promise<Omit<Application, "environments"> | null> => {
   const repository = context.get<IApplicationRepository>(
     APPLICATION_MONGO_DB_REPOSITORY,
   );
+
+  const application = await repository.getApplication({ appName, clerkId });
+  if (!application) {
+    throw new ServiceError(httpError.APP_CANT_DELETE);
+  }
 
   try {
     return repository.deleteApplication({ appName, clerkId });
